Moved the code related to local storage from components/Landing.js to helpers/storage.js, and put the call to storeRepo in state.js as suggested
